Modify Firestore AsyncQueue to be more robust
Firestore was becoming stuck because whatever operation was on `tail` seemed to never complete. Once in this state, all future writes would simply queue up in memory and never process, losing user data.

My theory is that disk errors (or hangs) are transient, and that trying again will probably work. This refactors AsyncQueue to add a few things:
1. Timeouts for all async operations
2. Automatic retry for each operation (up to a limit)
3. Logging when an operation fails

This is pretty hacky because attempting to link `@firebase/firestore` ended up including multiple versions of firebase, and I tweaked rollup to be faster and emit more readable code.